feat: add validation engine service
